backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
groq==0.4.1
sentence-transformers==5.1.1
huggingface_hub==0.35.3
faiss-cpu==1.7.4
chromadb==0.4.22
cachetools==5.3.2
aiohttp==3.9.1
